% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_iso2c.R
\docType{data}
\name{all_iso2c}
\alias{all_iso2c}
\title{Data frame of 253 ISO two letter country codes
Used interally when users request to download data for 'all' countries.}
\format{a data frame with 253 observations and 1 column}
\source{
From the countrycode package.
}
\usage{
all_iso2c
}
\description{
Data frame of 253 ISO two letter country codes
Used interally when users request to download data for 'all' countries.
}
\keyword{datasets}
