% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{imf_app_name}
\alias{imf_app_name}
\title{Set the IMF Application Name}
\usage{
imf_app_name(name = "imfr")
}
\arguments{
\item{name}{A string representing the application name. Default is "imfr".}
}
\value{
Invisible NULL
}
\description{
Set a unique application name to be used in requests to the IMF
API as a hidden environment variable
}
\details{
The `imf_app_name` function sets the application name that will be used in
the request header when making API calls to the IMF API. The IMF API has an
application-based rate limit of 50 requests per second, with the application
identified by the "user_agent" variable in the request header. The function
sets the application name by changing the `IMF_APP_NAME` hidden variable in
`.Renviron`. If this variable doesn't exist, `imf_app_name` will create it.
}
\examples{
imf_app_name("my_custom_app_name")

}
