% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf_data_list.R
\name{imf_data}
\alias{imf_data}
\title{Download an data from the IMF}
\usage{
imf_data(database_id, indicator, country, start = 2000, end = 2013)
}
\arguments{
\item{database_id}{character string database ID. Can be found using
\code{\link{imf_ids}}.}

\item{indicator}{character string indicator ID. Can be found using
\code{\link{imf_codes}}.}

\item{country}{character string or character vector of ISO two letter
country codes identifying the countries for which you would like to
download the data.See \url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}.}

\item{start}{time point for which you would like to start gathering the data.}

\item{end}{time point for which you would like to end gathering the data.}
}
\description{
Download an data from the IMF
}
\examples{
\dontrun{
# Download Real Effective Exchange Rate (CPI base) for the UK and China
real_ex <- imf_data(database_id = 'IFS', indicator = 'EREER_IX',
               country = c('CN', 'GB'))
}

}

