% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{imf_dataset}
\alias{imf_dataset}
\title{Download a data series from the IMF}
\usage{
imf_dataset(database_id, parameters, start_year, end_year)
}
\arguments{
\item{database_id}{character string. Database ID for database from which
you would like to request data. Can be found using
\code{\link{imf_databases}}.}

\item{parameters}{list of data frames providing input parameters for your
API request. Retrieve list of all possible input parameters using
\code{\link{imf_parameters}} and filter each data frame in the list to
reduce it to the inputs you want.}

\item{start_year}{integer four-digit year. Earliest year for which you would like
to request data.}

\item{end_year}{integer four-digit year. Latest year for which you would like to
request data.}

\item{return_raw}{logical. Whether to return the raw list returned by the API
instead of a cleaned-up data frame. This argument exists strictly for
purposes of backward compatibility with earlier versions of \pkg{imfr} and
will be discontinued in a future version.}

\item{print_url}{logical. Whether to print the URL used in the API call.
Can be useful for debugging.}

\item{times}{numeric. Maximum number of requests to attempt.}

\item{accounting_entry, activity, adjustment, age, classification, cofog_function, commodity, comp_method, composite_breakdown, counterpart_area, counterpart_sector, currency_denom, cust_breakdown, disability_status, education_lev, expenditure, financial_institution, flow_stock_entry, freq, functional_cat, gfs_sto, income_wealth_quantile, indicator, instr_asset, instrument_and_assets_classification, int_acc_item, maturity, occupation, prices, product, ref_area, ref_sector, reporting_type, series, sex, sto, summary_statistics, survey, transformation, type, unit_measure, urbanisation, valuation}{character vector. Use \code{imf_parameters} to identify which parameters to
use for requests from a given database and to see all valid input codes for
each parameter.}
}
\value{
Returns a tidy data frame with the data series.
}
\description{
Function to request data from a database through the IMF API endpoint.
}
\details{
Only the \code{database_id} argument is strictly required; all other
arguments are optional. If you provide a \code{database_id} without any other
arguments, the function will attempt to download the entire database.
However, many databases available through the API are too large to download
in their entirety, and your request will fail. Additional arguments to the
function act as filter parameters to reduce the size of the returned dataset.
For instance, supplying \code{c("A","M")} as the \code{freq} argument will
return all database observations of annual or monthly frequency, while
excluding all observations of quarterly frequency.

There are two ways to supply parameters for your API request. The optimal way
is to retrieve a list of data frames using \code{\link{imf_parameters}},
filter each data frame to retain only the parameters you want, and then
supply the modified list object to \code{imf_dataset} as its \code{parameters}
argument. However, users who are not comfortable modifying data frames in a
nested list may find it easier to instead supply one or more character
vectors as arguments, as in the example in the previous paragraph. (There are
a total of 44 possible parameters for making request from various databases
through the API, and each parameter uses unique input codes, which is why the
\code{parameters} list method simplifies things!) These two methods for
specifying parameters may not be combined. Only  \code{database_id},
\code{start_year}, \code{end_year}, \code{print_url}, and \code{times}
arguments may be used in combination with a \code{parameters} list object;
any other arguments will be ignored (and a warning thrown).
}
\examples{
# Retrieve "Assets (with Fund Record), National Currency" series from the
# Balance of Payments database using the list object method
params <- imf_parameters(database_id = 'BOP')
params$indicator <- filter(params$indicator,
                           description == 'Assets (with Fund Record), National Currency')
df <- imf_dataset(database_id = 'BOP', parameters = params)

# Retrieve "Assets (with Fund Record), National Currency" series from the
# Balance of Payments database using the character vector method
params <- imf_parameters(database_id = 'BOP')
indicator_code <- filter(params$indicator,
                           description == 'Assets (with Fund Record), National Currency')$input_code
df <- imf_dataset(database_id = 'BOP', indicator = indicator_code)

}
