% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{imf_ids}
\alias{imf_ids}
\title{List IMF database IDs}
\usage{
imf_ids(return_raw = FALSE, times = 3)
}
\arguments{
\item{return_raw}{logical. Whether to return the raw dataflow list or a
data frame with database IDs and names.}

\item{times}{numeric. Maximum number of requests to attempt.}
}
\value{
If \code{return_raw = FALSE} then a data frame with the database IDs
and descriptions is returned. If \code{return_raw = TRUE} then the raw
dataflow list is returned.
}
\description{
List IMF database IDs
}
\examples{
head(imf_ids())

}
