% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{imf_parameter_defs}
\alias{imf_parameter_defs}
\title{Get definitions of IMF API parameters}
\usage{
imf_parameter_defs(database_id, times = 3)
}
\arguments{
\item{database_id}{character string of a \code{database_id} from
\code{\link{imf_databases}}.}

\item{times}{numeric. Maximum number of API requests to attempt.}
}
\value{
Returns a data frame of input parameters used in making API requests
from a given IMF database, along with text descriptions or definitions of
those parameters. Useful in cases when parameter names returned by
\code{\link{imf_databases}} are not self-explanatory. (Note that the
usefulness of text descriptions can be uneven, depending on the database
design.)
}
\description{
Get text descriptions of input parameters used in making API
requests from a given IMF database
}
\examples{
# Get names and text descriptions of parameters used in IMF API calls to the
Balance of Payments database
imf_parameter_defs(database_id = 'BOP'))

}
