% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R, R/imfr-deprecated.R
\name{imf_ids}
\alias{imf_ids}
\alias{imf_codelist}
\alias{imf_codes}
\alias{imf_data}
\alias{imf_metastructure}
\alias{imf_metadata}
\alias{imfr-deprecated}
\title{List IMF database IDs}
\usage{
imf_ids(return_raw = FALSE, times = 3)

imf_codelist(database_id, return_raw = FALSE, times = 3)

imf_codes(codelist, return_raw = FALSE, times = 3)

imf_data(
  database_id,
  indicator,
  country = "all",
  start = 2000,
  end = format(Sys.Date(), "\%Y"),
  freq = "A",
  return_raw = FALSE,
  print_url = FALSE,
  times = 3
)

imf_metastructure(database_id, return_raw = FALSE, times = 3)

imf_metadata(database_id, URL, times = 3, ...)
}
\arguments{
\item{return_raw}{logical. Whether to return the raw metadata
structure list or a data frame with codelist codes and descriptions.}

\item{times}{numeric. Maximum number of requests to attempt.}

\item{database_id}{character string. database_id to request the header for.
Can be found using \code{\link{imf_databases}}.}

\item{codelist}{character string of a \code{codelist} from
\code{\link{imf_codelist}}.}

\item{indicator}{character string or character vector of indicator IDs.
These can be found using \code{\link{imf_codes}}.}

\item{country}{character string or character vector of ISO two letter
country codes identifying the countries for which you would like to
download the data.See \url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}.
If \code{country = 'all'} then \code{imf_data} will attempt to download
all available countries.}

\item{start}{year for which you would like to start gathering the data.}

\item{end}{year for which you would like to end gathering the data.}

\item{freq}{character string indicating the series frequency. With
\code{'A'} for annual, \code{'Q'} for quarterly, and \code{'M'} for monthly.}

\item{print_url}{logical. Whether to print the URL used in the API call.
Can be useful for debugging.}

\item{URL}{character string. Used internally by \code{imf_databases} to
request header by request URL rather than by database_id.}
}
\value{
If \code{return_raw = FALSE} then a data frame with the database IDs
and descriptions is returned. If \code{return_raw = TRUE} then the raw
dataflow list is returned.

If \code{return_raw = FALSE} then a data frame with the codelist IDs
and descriptions is returned. If \code{return_raw = TRUE} then the raw
data structure list is returned.

If \code{return_raw = FALSE} then a data frame with the codes
and descriptions is returned. If \code{return_raw = TRUE} then the raw
code list is returned.

If \code{return_raw = FALSE} then a data frame with just the
requested data series are returned. If \code{return_raw = TRUE} then the raw
data list is returned. This can include additional information about the
series.

If \code{return_raw = FALSE} then a data frame with the codelist IDs
and descriptions is returned. If \code{return_raw = TRUE} then the raw
data structure list is returned.
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\section{\code{imf_ids}}{

For \code{imf_ids}, use \code{\link{imf_databases}}.
}

\section{\code{imf_codelist}}{

For \code{imf_codelist}, use \code{\link{imf_parameters}}.
}

\section{\code{imf_codes}}{

For \code{imf_codes}, use \code{\link{imf_parameters}}.
}

\section{\code{imf_data}}{

For \code{imf_data}, use \code{\link{imf_dataset}}.
}

\section{\code{imf_metastructure}}{

Function \code{imf_metastructure} will be discontinued in a future version
for lack of evident use cases.
}

\section{\code{imf_metadata}}{

Function \code{imf_metadata} is deprecated and will be discontinued in a
future version. Use \code{imf_dataset(include_metadata = TRUE)} instead.
}

\examples{
suppressWarnings(databases <- imf_ids())

#' # Find Primary Commodity Price System database data structure
suppressWarnings(cl <- imf_codelist(database_id = 'PCPS'))

# Retrieve indicators from BOP database
suppressWarnings(codes <- imf_codes(codelist = 'CL_INDICATOR_BOP'))

# Download Real Effective Exchange Rate (CPI base) for China at an annual
# frequency since 2018
suppressWarnings(df <- imf_data(database_id = 'IFS', indicator = 'EREER_IX',
                 country = 'CN', freq = 'A',start = 2018))

# Find Primary Commodity Price System database data structure
suppressWarnings(metastruc <- imf_metastructure(database_id = 'PCPS'))

# Find Primary Commodity Price System database metadata
suppressWarnings(metadata <- imf_metadata(database_id = 'PCPS'))

}
\keyword{internal}
