% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{set_imf_use_cache}
\alias{set_imf_use_cache}
\title{Set the IMF cache usage option}
\usage{
set_imf_use_cache(use_cache)
}
\arguments{
\item{use_cache}{A boolean value indicating whether the cache should be used.}
}
\value{
The updated `imf_use_cache` option value.
}
\description{
This function allows users to set the `imf_use_cache` option, which determines
whether the cache should be used when fetching data from the International Monetary Fund (IMF).
}
\examples{
set_imf_use_cache(TRUE)
set_imf_use_cache(FALSE)
}
