% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{set_imf_wait_time}
\alias{set_imf_wait_time}
\title{Set the IMF wait time}
\usage{
set_imf_wait_time(new_wait_time)
}
\arguments{
\item{new_wait_time}{Numeric value representing the new wait time (in seconds) between API requests.}
}
\value{
None.
}
\description{
This function allows you to modify the mandatory wait time between API calls with the imfr package.
Wait time should be a numeric value between 0 and 10 (in seconds). It is not recommended to use
wait times greater than 5 seconds, as they will significantly slow runtimes. Wait times greater
than 10 seconds are not allowed. The function sets the wait time by changing the `imf_wait time` R
option. This option will not persist across sessions. To set a persistent app name, add this
function call to your .RProfile.
}
\examples{
# Usage
library(imfr)
set_imf_wait_time(2.0)  # Change the wait time to 2 seconds
}
