% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{imf_databases}
\alias{imf_databases}
\title{List IMF database IDs and descriptions}
\usage{
imf_databases(times = 3)
}
\arguments{
\item{times}{numeric. Maximum number of API requests to attempt.}

\item{return_raw}{logical. Whether to return the raw dataflow list or a
data frame with database IDs and names.}
}
\value{
Returns a data frame with \code{database_id} and text
\code{description} for each database available through the IMF API endpoint.
}
\description{
List IMF database IDs and descriptions
}
\examples{
# Return first 6 IMF database IDs and descriptions
head(imf_databases())
\dontrun{
# Open a viewing pane with the full list of IMF database IDs and descriptions
View(imf_databases())
}

}
